Rename android package from RNPackages -> RNShowcase
