+++ android/app/src/main/java/com/rnshowcase/MainActivity.java
@@ -1,4 +1,4 @@
-package com.rnpackages;
+package com.rnshowcase;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "RNPackages";
+        return "RNShowcase";
     }
 }
