+++ android/app/src/main/java/com/rnshowcase/MainApplication.java
@@ -1,14 +1,8 @@
-package com.rnpackages;
+package com.rnshowcase;
 
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
-import com.microsoft.codepush.react.CodePush;
-import com.robinpowered.react.Intercom.IntercomPackage;
-import com.jadsonlourenco.RNShakeEvent.RNShakeEventPackage;
-import com.horcrux.svg.SvgPackage;
-import com.brentvatne.react.ReactVideoPackage;
-import com.rnuxcam.rnuxcam.UXCamPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -20,12 +14,6 @@ import java.util.List;
 public class MainApplication extends Application implements ReactApplication {
 
   private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
-
-    @Override
-    protected String getJSBundleFile() {
-      return CodePush.getJSBundleFile();
-    }
-
     @Override
     public boolean getUseDeveloperSupport() {
       return BuildConfig.DEBUG;
@@ -34,13 +22,7 @@ public class MainApplication extends Application implements ReactApplication {
     @Override
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
-          new MainReactPackage(),
-            new CodePush(null, getApplicationContext(), BuildConfig.DEBUG),
-            new IntercomPackage(),
-            new RNShakeEventPackage(),
-            new SvgPackage(),
-            new ReactVideoPackage(),
-            new UXCamPackage()
+          new MainReactPackage()
       );
     }
   };
